backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1
pillow==10.2.0
easyocr==1.7.1
pytesseract==0.3.10
PyMuPDF==1.23.26
opencv-python-headless==4.9.0.80
scikit-learn==1.4.0
numpy==1.26.4
rake-nltk==1.0.6
nltk==3.8.1
joblib==1.3.2
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.26.0
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.26.0
psycopg2-binary==2.9.9
